Initial dump into the repo.
This is not the true beginning files for the site. I started off with a nanoc site  doing the SSG game.. but the yaml frontmatter for my pages got outta wack and it wasn't really progressing like I wanted.. so it was scrapped and this filler page was thrown together in a hot mess of a minute. 

:spaceship: the space Cowboy wants to be found. His logs are still being decoded.